refactor(svelte): split pure semantic-keys helpers by concern
semantic-keys.ts mixed source identity, data/spec epoch tokens,
resolveSemanticKeys, and the candidate projection cache. Extract four
sibling modules and retarget internal imports; drop the mixed file.

Closes #1163.
