+++ packages/svelte/tests/legend/focus-state.integration.test.ts
@@ -17,7 +17,7 @@ import { createLegendEntryKeyIndex } from "../../src/lib/legend/entry-key-index.
 import { createLegendFocusState } from "../../src/lib/legend/focus-state.svelte.js";
 import type { InteractiveLegendEntry, LegendEntryIdentity } from "../../src/lib/legend/focus.js";
 import { createPlotRuntime } from "../../src/lib/runtime/runtime.svelte.js";
-import { createSourceIdentityTracker } from "../../src/lib/runtime/semantic-keys.js";
+import { createSourceIdentityTracker } from "../../src/lib/runtime/semantic-source-identity.js";
 import { createSemanticKeyService } from "../../src/lib/runtime/semantic-keys.svelte.js";
 import { withFlushedEffectRoot } from "../helpers/effect-root.svelte.js";
 import { createReactiveRuntimeDeps } from "../helpers/runtime-deps.svelte.js";
